Many changes e.g. design reform (incl. proper mobile support), permission system

+++ src/pages/delete.js
@@ -1,39 +1,54 @@
-module.exports = (req, res, username, users, pages, recentchanges, history) =>
+module.exports = (req, res, username, users, pages, recentchanges, history, perm) =>
 {
     //username parameter: reserved for history
     //todo: ACL
-    pages.findOne({where: {title: req.params.name}}).then(page =>
+    if (username === undefined)
     {
-        var doneby = req.session.username
-        if (doneby === undefined) doneby = req.connection.remoteAddress
-        if (page) //if page exists
+        require(global.path + '/error.js')(req, res, null, 'Please login.', '/login', 'the login page')
+        return
+    }
+    perm.findOne({where: {username: username, perm: 'grant'}}).then(p =>
+    {
+        if (p)
         {
-            const oldLength = page.content.length
-            pages.destroy({where: {title: req.params.name}})
-            history.destroy({where: {page: req.params.name}})
-            recentchanges.create(
+            pages.findOne({where: {title: req.params.name}}).then(page =>
             {
-                page: req.params.name,
-                doneBy: doneby,
-                bytechange: -oldLength,
-                comment: req.body.comment,
-                type: 'delete'
+                var doneby = req.session.username
+                if (doneby === undefined) doneby = req.connection.remoteAddress
+                if (page) //if page exists
+                {
+                    const oldLength = page.content.length
+                    pages.destroy({where: {title: req.params.name}})
+                    history.destroy({where: {page: req.params.name}})
+                    recentchanges.create(
+                    {
+                        page: req.params.name,
+                        doneBy: doneby,
+                        bytechange: -oldLength,
+                        comment: req.body.comment,
+                        type: 'delete'
+                    })
+                    history.create(
+                    {
+                        page: req.params.name,
+                        rev: 0,
+                        bytechange: -oldLength,
+                        editedby: doneby,
+                        comment: req.body.comment,
+                        type: 'delete'
+                    })
+                    res.redirect('/')
+                }
+                else
+                {
+                    //error!
+                    require(global.path + '/error.js')(req, res, null, 'No such page.', '/', 'the main page')
+                }
             })
-            history.create(
-            {
-                page: req.params.name,
-                rev: 0,
-                bytechange: -oldLength,
-                editedby: doneby,
-                comment: req.body.comment,
-                type: 'delete'
-            })
-            res.redirect('/')
         }
         else
         {
-            //error!
-            require(global.path + '/error.js')(req, res, null, 'No such page.', '/', 'the main page')
+            require(global.path + '/error.js')(req, res, null, 'You do not have permission to delete page.', '/login', 'the login page')
         }
     })
 }
